+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Remove multi-threaded account seed generation due to single-threaded generation being faster (#1061).
 - Made `AccountIdError` public (#1067).
 - Made `BasicFungibleFaucet::MAX_DECIMALS` public (#1063).
+- [BREAKING] Removed `miden-tx-prover` crate and created `miden-proving-service` and `miden-remote-provers` (#1047).
 
 ## 0.6.2 (2024-11-20)